# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HighLevel PayTR Integration API",
    "description": "API documentation for HighLevel PayTR payment integration. This API enables Turkish payment processing through PayTR within the HighLevel CRM platform.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/payments/page": {
      "get": {
        "tags": ["Payment"],
        "summary": "Payment iframe page",
        "description": "Returns the payment page with embedded PayTR iframe",
        "operationId": "paymentPage",
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float",
              "example": 100.50
            }
          },
          {
            "name": "email",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "email",
              "example": "customer@example.com"
            }
          },
          {
            "name": "transactionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "txn_123456789"
            }
          },
          {
            "name": "locationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "loc_123"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "TRY",
              "enum": ["TRY", "USD", "EUR"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment page HTML"
          },
          "400": {
            "description": "Invalid parameters"
          },
          "401": {
            "description": "Invalid account"
          }
        }
      }
    },
...
    "/payments/error": {
      "get": {
        "tags": ["Payment"],
        "summary": "Payment error page",
        "description": "Redirect page shown after failed payment",
        "operationId": "paymentError",
        "parameters": [
          {
            "name": "merchant_oid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Error page HTML"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "LocationHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Location-Id",
        "description": "HighLevel location ID for authentication"
      }
    },
...
  }
}